Lib file renamed
