+++ bin-src/lib/comparePackageManifests.ts
